move: move traversal classes
